+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - [commands] Fixes MapBasedParser tabCompletion index
 - [commands] Added error logging for cases where subcommand methods are non-public or when variables are misused.
 - [commands] Added `MethodUtils#getAllMethods(Class)` for getting all methods of a class including inherited & public & non-public
+- [commands] Fixes bug with `ExecptionHandle` ignoring value
+- [commands] Removed `ExecptionArgumentMapper` in favor of `ArgumentMapper`
   
 ## 0.0.37
 - [commands] Enhanced command handling with improved serialization of subcommands